Make autopostgresqlbackup a bit more portable
GNU `stat` is not very portable. The less obvious, more portable and
posix-friendly way to get the bytecount of a file is to use `wc(1)` -
`wc -c <
file`. This will work across all the Linuxes/unixes/BSDs.

`hostname` is also not very portable. Especially not with the
`--fqdn`-flag. Nn fact, the man-page of
[hostname(1)](https://manpages.debian.org/bookworm/hostname/hostname.1.en.html)
kinda warns against using it.

The POSIX-friendly way of getting the hostname is `uname -n`, which will
return the "node name". I believe `uname -n` uses the same
`gethostname(2)` which `hostname` uses. FQDN can be set either according
to hostname(1), or by hostnamectl(1) on systemd systems.